fix lesson1

+++ lesson1/triangle_area.rb
@@ -1,8 +1,8 @@
 print 'Input triangle base: '
-base = gets.chomp.to_i
+base = gets.chomp.to_f
 
 print 'Input triangle height: '
-height = gets.chomp.to_i
+height = gets.chomp.to_f
 
 puts '-----------------------------------'
 puts "Tiangle area is #{base * height * 0.5}."
